+++ src/cli/commands/mod.rs
@@ -1,3 +1,4 @@
+pub mod ble;
 pub mod dfu;
 pub mod list;
 pub mod loglevel;